requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.3
python-socketio==5.8.0
python-engineio==4.4.1
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
bidict==0.22.1
Pillow==9.5.0
python-dotenv==1.0.0 